requirements.txt
Flask==3.1.3
python-dotenv==1.2.2
botbuilder-core==4.17.1
botbuilder-schema==4.17.1
playwright==1.59.0
requests==2.33.1
aiohttp==3.13.5
